Fix breaks in Python 3.5 due to pathlib (#12)
Because Black is Python 3.6 only, our CI builds were silently
upgrading Python to 3.6. So our 3.5 builds weren't actually 3.5.
Removing black from the `requirements.txt` and install it 
manually on `.travis.yml` when running the checks works.
Leave it in the `environment.yml` so that devs will get it installed.

Breaks on 3.5 are related to mixing `os.path` and `pathlib`.
Use pathlib in most places now and include `str` conversions
because 3.5 doesn't support passing Path to `open` etc.