Check if data dir is writable and warn the user
`pooch.create` now only checks for a PermissionError when trying to
create the data directory and warns the user to set the environment
variable if the error occurs. Also checks if we can create a tempfile in
the directory if it already exists and issues the same warning.
The exception is not passed on so that modules that call `create` won't
break at import time. Instead, the failure will happen when
`Pooch.fetch` tries to create the directory or write to a file if it
exists.

Fixes #22