Allow custom download URLs for individual files
Custom URLs are specified by the new `urls` optional argument for
`pooch.create` and `pooch.Pooch`. It is a dictionary matching file names
to download URLs. If a file is in `Pooch.urls`, it will be downloaded
from there instead of the `base_url`. Custom URLs can be loaded from
registry files as well by appending the URL to the end of a line.

This will be useful if your project downloads data from a remote source
that you don't control. Otherwise, the only choice would be to duplicate
the data in your repository (which would be bad if the files are large).

Fixes #4