Use appdirs to get the cache directory (#45)
Rely on this is tiny module (no dependencies, one file, 600 lines) to do the right thing
instead of implementing it ourselves. It follows the XDG specs for *NIX, but also
correctly reads the location for Windows from the registry.

Adds `appdirs` as a dependency. Breaks compatibility by removing the `platform` argument
of `os_cache` since it's no longer used. This should not be an issue since it was only
there for testing purposes.

Fixes #43