Add post-download hook to Pooch.fetch (#59)
Add a new argument `hook` to `Pooch.fetch` to pass in a function that is
executed right before returning and can overwrite the file path that is
returned to the user. Example usage would be to unzip an archive only
when downloading the file and return the path to the unzipped file. Will
be very useful to perform unpacking operations on larger files that can
be time consuming and we only want to do once.

Fixes #35