Add support for file objects to Pooch.load_registry() (#117)
This makes it possible to pass a file object to `Pooch.load_registry()` 
(opened either in binary or text mode).

My main reason for this is to support passing `pkg_resources.resource_stream()` 
from `setuptools`. This is the preferred way of using data files from the package.
It will allow your package to be "zip-safe".