Improve project description in the README (#124)
In preparation for our JOSS paper (#112), update the description of the
README to highlight the advantages and main features of Pooch for
package authors and scientists. Move the example down (it's not really
tl;dr if it's longer than the About). Also use the general DOI for the
package for citation instead of the version specific one so we don't
have to update it every time.