Use file name in exception for hash mismatch (#128)
When the hash of the downloaded file doesn't match the one in the
registry, we raise an exception informing the hashes and the name of the
downloaded temporary file. But the temp file is deleted and its name has
nothing to do with the file in the registry. Instead, inform the actual
file name.

Fixes #96