Refactor code for checking the file hash
Move code from methods to 2 dedicated functions: one to check the hash
algorithm and the other to check that the hash matches a known hash.
The hash algorithm defaults to sha256 and hash checking functions knows
to ignore the algorithm part of the hash when checking (if needed).
This simplifies the class by removing the need to add algorithms to all
hashes (retaining the legacy behaviour) and removing some duplicated
code for calculating hashes in the Pooch class.