Move downloader choice to a separate function
The code that chooses the appropriate downloader for the given url
doesn't depend on anything else in `pooch/core.py`. Move it to
`pooch/downloaders.py` to test it separately from the `Pooch` class and
avoid importing downloaders at all in `pooch/core.py`. Move the
downloader specific tests to a separate file as well so it matches the
tests for processors better (in a separate file instead of in the core
tests).