Don't use temp file name in hash mismatch error
Fixes a regression introduced in #157 where we started including the
name of the temporary download file in the exception again. However,
since the download is deleted in case of hash mismatch, it's also not
very useful to print the name of a file that doesn't exist. Instead, I
included the download URL in the exception so that a user can download
for themselves and check.