Add the retrive function to fetch single files (#152)
Add the function `retrieve` for downloading single data files. 
This is much more convenient than setting up a `Pooch`. 
It automatically selects a unique file name and saves files to 
a "pooch" cache (as determined by `os_cache`). Allows 
passing in `known_hash=None` to bypass the hash checks. 
In this case, will print a message with the SHA256 hash of 
the downloaded file and advising users to set `known_hash` 
to this value.

Fixes #147