Fix FTP availability check when file in directory
If the data file is not in the base directory, the Pooch.is_available
test was broken since we were checking for the full path in `ftp.nlst`
instead of just the file name. Using the UNAVCO ftp server revealed this
(ftp://data-out.unavco.org/). The change was made since the one we were
using before started refusing connections on CI.