Test FTP downloads using a local test server (#192)
Using the `pytest-localftpserver` plugin, we can test FTP downloads running on a local server. 
The nice thing is that the plugin just provides a fixture that can access the server so we don't 
have to do much. Added a context manager to move test files to the local server and clean up 
afterwards. The availability test is still relying on a remote server because the local server runs 
on a random port and we don't have a way to control that on `Pooch.is_available` yet.

Fixes #190