Set up GitHub Actions for testing and linting (#194)
Use Github Actions as a replacement for Azure and TravisCI non-deploy builds. 
Runs faster and was easier to configure. Not using the fatiando/continuous-integration 
scripts but should migrate as soon as those are ready. A lot of those scripts can be
replaced with available Actions. Set up two workflows: 1 for linting and 1 for testing.
The linter workflow runs black, flake8, and pylint in separate jobs we can see which 
tool passes/fails. The tests are run on linux/mac/windows with Python >=3.5. Docs
are uploaded as artefacts for later deployment and visual inspection.