Push documentation to GitHub Pages using Actions (#198)
Replace the TravisCI deployment of documentation pages with Actions.
Had to rebuild the docs in the `deploy` workflow since we can't share build
artefacts between workflows at the moment. Implement the script to push
the updated docs here instead of relying on the `continuous-integration` 
repository. This eliminates the need for TravisCI, so also remove the 
configuration file for it.