Add warning about dropping Python 3.5 to the docs
Update the README message to say that 1.2.0 is the last to support 3.5.
Add a compatibility table to the Installing page. Some other
improvements to the install instructions (move dependencies to the
bottom and say that they should be installed automatically, add a note
that the testing step is optional, link to Miniconda as well, using
`python -m pip` instead of `pip`, use `git+https://...` instead of the
zip archive to get correct version numbers from versioneer).