Use nox to manage development automation
The [Nox](https://nox.thea.codes) tool allows us to specify commands
that are run inside virtual environments automatically. This lets us
run tests on multiple Python version, using conda or pip, with or
without optional dependencies, etc, locally instead of just on CI. Plus,
the configuration is written in Python (`noxfile.py`) so it's
cross-platform and easy to enhance (unlike our Makefile). The
configuration here uses requirements files in `env` to create the needed
environments for different tasks. Automation includes cleaning generated
files, producing source and wheel archives, building the docs, and
creating a sandbox conda environment to test and prototype. This would
completely replace the main `Makefile` and `doc/Makefile` (kept for
compatibility but will be removed in the future).