Add option to retry downloads if they fail (#215)
The new `retry_if_failed` option to `create` and `Pooch` allows retrying
the download the specified number of times in case of failures due to
hash mismatches (coming from Pooch) or network issues (coming from
`requests`). This is useful for running downloads on CI that tend to
fail sporadically. Waits a period of time between consecutive downloads
starting with 1s and increasing up to 10s in 1s increments.
Fixes #205