Refactor the docs into separate tutorials (#237)
The docs were organized by level (beginner to advanced), which is good if 
reading it from start to finish. But that format is not great when going back to 
try to find out how to do something. Here, I break apart the larger pages into
separate tutorials based on topic. The beginner tutorial is moved to "Getting
started" and split into "Fetching multiple files" and "Managing a package's
data". Also specify the API pages as "List of functions and classes (API)" 
since novices are likely not used to the term API.