Redesign the front page of the docs (#238)
Remove the wall of text and make clear boxes for getting started, asking for help, 
API reference, and citing. Include the full table of contents underneath to help with 
navigation. Move the text about different use cases to a "Why Pooch?" page. 
Introduces the `sphinx-panels` extension to use cards, columns, and tabs.