Add support for figshare and Zenodo DOIs as URLs (#241)
Adds the `DOIDownloader` class, which uses the public API of each reposiroty 
to get the actual HTTP download link, which is then passed to `HTTPDownloader`. 
Defines the new URL protocol `doi:{DOI}/{file name}` so that everything can 
happen automatically in `Pooch` and `retrieve`. 
Test data are stored in Zenodo (https://doi.org/10.5281/zenodo.4924875/) and
figshare (https://doi.org/10.6084/m9.figshare.14763051).
One caveat is that we can only download 1 file from the archive and the file name 
must be specified in the URL. This is because our downloaders are all designed 
to fetch a single file and save it to a given location. Downloading multiple files 
would require extensive refactoring of large portions of code. 

Fixes #177