Run CI jobs on Python version extremes
We don't need to run on all versions, just the two end members. Chances
of something not working on a version between the lowest and highest
version supported are pretty much zero. Instead, run with optional
dependencies on both end members as well.

Took this chance to update the CI configuration with latest changes from
the other libraries.