Fix bug when unpacking a subfolder from an archive (#266)
Our unpacking processors (`Untar` and `Unzip`) aren't handling `members` that
are a whole subfolder correctly. The elements of `members` are being treated as
files regardless of what they are. Add a test for use case (no change to our
test data required) and refactor the unpacking test to make it easier to extend
in the future.