Add option to disable updates (Fixes #290)
This gives users the option, when creating a Pooch, to disallow updating
local copies of files. The intended use case for this is for development
checkouts/CI systems where the local copies are canonical and any need
to update likely reflects a problem with the registry (like and out of
date hash).