Rewrite archive extraction with members argument
So far, the logic whether extraction needs to be run depended on the target
directory existing on the filesystem. This made it impossible to extract multiple
times from the same archive with differing members. The commit changes the
behaviour to checking whether all required files already present.

In the process, I also had to normalize the members parameter to not
have leading "./".