Make archive extraction always take members into account (#316)
The logic whether extraction needs to be run depended on the target directory
existing on the file system. This made it impossible to extract multiple times
from the same archive with differing members. Change the behaviour to checking
whether all required files already present. In the process, also normalize the
members parameter to not have leading `./`.