Refactor Pooch.is_available to use downloaders (#322)
The method checks if a file is available without downloading it. It used
an implementation independent of the downloaders and only supported
simple HTTP and FTP. To allow for more flexibility, allow the function
to take a *downloader* argument. Downloaders had to be modified to
include an optional check_only=False argument to do the checking and
return a bool instead of None. Pooch.is_available raises an exception if
the argument isn't supported by a given downloader, so we can retain
backwards compatibility. Allowed refactoring the FTP is_available test
to use the pytest-localftpserver instead of relying on a public server
somewhere (the UNAVCO one we used was decommissioned).