Populate registry using DOI-based data repositories (#325)
Add a new `Pooch.load_registry_from_doi` method that populates the
registry of the `Pooch` instance by making an API request to the DOI
repository and getting all the available files and their corresponding
hashes. Add `DataRepository` classes the ability to cache the API
responses in order to minimize the number of times we need to request
it. Refactor the `DataverseRepository` class: create a new `classmethod`
that can perform the API request before or after the initialization of
the object, reuse this method when any API request is needed to be
performed. Add test functions for this new feature. Add small example in
the User Guide on how to use this new method.

Co-authored-by: Santiago Soler <santisoler@fastmail.com>