Support Zenodo urls with slashes in the filename (#340)
Allow `DOIDownloader` to work with files uploaded to Zenodo that have a forward slash "/" in their name. These special cases are generated by the Zenodo-GitHub integration service, where the first portion of the filename matches the GitHub handle of the repository owned and the last portion matches the repository name. Now the `parse_url` function handles Zenodo dois under a special case: the netloc will be formed with the first two "parts" (`prefix/suffix`), while the path will be formed by everything that follows the second slash.