Fix bug: Figshare downloaders fetch correct version
Fix bug that made Figshare downloaders to fetch always the latest
version. Now the `api_response` property parses the DOI to get the
desired version of the repository and makes use of the Figshare API to
get the list of files available for that particular version. If no
version is specified through the DOI, the last version will be fetched,
keeping backward compatibility and the default behaviour of Zenodo.
Add test functions.