Fix bug: Figshare downloaders fetch correct version (#343)
Fix bug that made Figshare downloaders to fetch always the latest
version. Now the `api_response` property of `FigshareRepository` parses
the DOI to get the desired version of the repository and makes use of
the Figshare API to get the list of available files for that particular
version. If no version is specified through the DOI, the last version
will be fetched, keeping backward compatibility and the default
behaviour of Zenodo. Add test functions.