Only create local data directories if necessary.
Creating a local data directory with Pooch checks that the local
directory is writeable. This is important if downloading data, but if
the files already exist is unnecessary.

This came up because we would like to store data for Pooch in readable
but non-writable directories. In particular, the directories in which
our build system installs Python are not writable.