Only create local data directories if necessary (#370)
Creating a local data directory with Pooch checks that the local
directory is writable. This is important if downloading data, but if
the files already exist is unnecessary. Doing so makes it impossible
to store data for Pooch in readable but non-writable directories. 
In particular, the directories in which some build systems install
Python may not be writable.