Speed up import time by lazy loading requests (#328)
Pooch is often used as an import addon but not as a core feature. 
As such it is a little silly to expand the import time of dependent 
packages by a lot. While the gain here is small, it is not 
insignificant when considering that multiple packages are using
Pooch and would any small amount of time would add up.