Fix bug: update Zenodo downloader for new API
Update the Zenodo downloader to work with the new Zenodo API. Use the
`filename` key for getting the filenames of all files in a repository.
Update the generated download url based on heuristics: the link present
in the API reference doesn't work at the moment). Update the method that
populates the registry: use the new `filename` key and specify that the
checksums are now md5.