Fix bug: add support for old and new Zenodo APIs (#375)
Allow the Zenodo downloader class to work both with the previous Zenodo
API and the new API that is running since their migration to InvenioRDM
on 2023-10-13. The downloader class `ZenodoRepository` determines which
API version is interacting with and defines the download url and
populate the registry based on that. Test if `ZenodoRepository`
correctly supports both the old and new Zenodo APIs. Use
`pytest-httpserver` to run a local server that returns API response that
mimics the Zenodo ones. Define a `base_api_url` class attribute for
`ZenodoRepository` so we can override it in tests.