Ensure all archive members are unpacked on a second call (#365)
After running Untar/Unzip with a single member passed and then running
again with `members=None` we'd only check that the unzipped folder 
exists and then return only the single member. This ensures that we 
return all of the requested files by checking that they exist in the unpacked
folder. 
Co-authored-by: Leonardo Uieda <leo@uieda.com>