Fix previous implementation
Restructure how the two APIs were being supported: the previous
implementation would raise an error when "md5" and "checksum" were not
available in the response, and not check for any other algorithms.
Correctly parse the checksum algorithm that is available in the new
Dataverse API: we need to convert "SHA-1" to "sha1" to name one.