Add mypy to CI job and type hints for one class. (#404)
Add type hints to `pooch/core.py` and create a new `typing` submodule
for custom type classes, and add it to the API Reference. Run `mypy` on
CI to perform type checks, and create new targets in the `Makefile`.
Extend the list of dependencies required to run the type checks.

---------

Co-authored-by: Santiago Soler <santisoler@fastmail.com>