Only run tests with network access on some CI jobs
Our CI is continuously hitting some external network providers which is
causing some of them (mostly figshare for now) to block our traffic.
This means that our CI fails randomly and it's annoying. Only run
network tests on jobs with the latest Python and optional dependencies
installed to try to mitigate this.